+++ src/redturtle/volto/interfaces.py
@@ -9,6 +9,17 @@ from plone.volto.interfaces import IPloneVoltoCoreLayer
 from redturtle.volto import _
 from zope.interface import Interface
 from zope.schema import Bool
+from zope.schema import SourceText
+
+import json
+
+
+# Default JSON configuration
+DEFAULT_RANKING_CONFIG = [
+    {"index": "Subject", "value": "__TERM__", "weight": 16},
+    {"index": "Title", "value": "__TERM__", "weight": 8},
+    {"index": "Description", "value": "__TERM__", "weight": 6},
+]
 
 
 class IRedturtleVoltoLayer(IDefaultBrowserLayer, IPloneVoltoCoreLayer):
@@ -30,6 +41,17 @@ class IRedTurtleVoltoSettings(Interface):
             default="If enabled, a custom ranking for SearchableText searches will be used.",
         ),
         default=False,
+        required=False,
+    )
+    advanced_query_ranking_rules = SourceText(
+        title=_(
+            "advanced_query_ranking_rules_label", default="AdvancedQuery Ranking rules"
+        ),
+        description=_(
+            "advanced_query_ranking_rules_help",
+            default="List of AdvancedQuery ranking rules. Use '__TERM__' for current search term.",
+        ),
+        default=json.dumps(DEFAULT_RANKING_CONFIG, indent=2),
         required=False,
     )